aider: Parse go.mod file content to extract dependencies.
USER: so i'd like a function that takes a go.mod file from the go language and returns a list of the dependencies for the go program it describes

ASSISTANT: I will create a function called `parse_go_mod_dependencies` that will take the content of a `go.mod` file as a string and return a list of dependencies.

Here are the diffs for the changes to `bin/process-go-dependencies.py`:

```diff
--- bin/process-go-dependencies.py
+++ bin/process-go-dependencies.py
@@ ... @@
 import requests

+def parse_go_mod_dependencies(go_mod_content):
+    """
+    Parses the go.mod file content and returns a list of dependencies.
+
+    :param go_mod_content: Content of a go.mod file
+    :return: List of dependencies
+    """
+    dependencies = []
+    lines = go_mod_content.split('\n')
+    inside_require_block = False
+    for line in lines:
+        line = line.strip()
+        if line.startswith('require ('):
+            inside_require_block = True
+            continue
+        if inside_require_block:
+            if line.startswith(')'):
+                inside_require_block = False
+                continue
+            if line:
+                dependencies.append(line.split(' ')[0])
+    return dependencies
+
 url = "https://raw.githubusercontent.com/filecoin-project/lotus/master/go.mod"
 response = requests.get(url).content
+dependencies = parse_go_mod_dependencies(response.decode('utf-8'))
+
+# Example usage: Print the list of dependencies
+for dep in dependencies:
+    print(dep)
```

This code will fetch the `go.mod` file from the specified URL, parse it to extract the dependencies, and then print them out.