aider: Creates a dict mapping dependencies to tags
USER: go through all the parsed dependencies, and for each one, create an entry in a dict. the dict key will be the dependency name, and the value will be any and all tags whose values the dependency name contains

ASSISTANT: Here are the diffs for the changes to `bin/process-go-dependencies.py`:

```diff
--- bin/process-go-dependencies.py
+++ bin/process-go-dependencies.py
@@ ... @@
 def parse_go_mod_dependencies(go_mod_content):
     """
     Parses the go.mod file content and returns a list of dependencies.
@@ ... @@
 tags = {
         "corporate_sponsor" : ['uber', 'google', 'gogo', 'zondax', 'hashicorp'],
         "ecosystem_projects" : ['filecoin-project','ipfs','libp2p', 'multiformats', 'ipld'],
         "batteries_included" : ['golang.org', 'gotest.tools', 'golang' ]
         }

+dependency_tags = {}
+
 url = "https://raw.githubusercontent.com/filecoin-project/lotus/master/go.mod"
 response = requests.get(url).content
 dependencies = parse_go_mod_dependencies(response.decode('utf-8'))
+
+for dependency in dependencies:
+    dependency_tags[dependency] = []
+    for tag, tag_values in tags.items():
+        if any(tag_value in dependency for tag_value in tag_values):
+            dependency_tags[dependency].append(tag)
```

This code will create a dictionary called `dependency_tags` where each key is a dependency name and the value is a list of tags that apply to that dependency.